+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # Braintree iOS SDK Release Notes
 
-## Unreleased
+## 4.29.0 (2019-09-19)
 
 * Fix issue when returning from the Venmo app on iOS13
 * Fix crash and return error when `threeDSecureRequest.amount` is `nil` and 3DS v1 is requested
 
-## 4.28.0 (2019-09-05) 
+## 4.28.0 (2019-09-05)
 
 * Add ability to customize UI for 3D Secure challenge views
 * Add authentication and lookup transaction status information to BTThreeDSecureInfo